app/tools/qdrant_retriever.py
from qdrant_client import QdrantClient
from langchain_qdrant import QdrantVectorStore
from langchain_core.embeddings import Embeddings
from app.core.config import settings

# Since we need an embedder to search Qdrant, we'll create a dummy one if no real one is specified,
# though Anthropic doesn't have a public embedding API directly in Langchain yet for some versions,
# or we can use another open source one. For this skeleton, we'll define a placeholder Embeddings class.

class DummyEmbeddings(Embeddings):
    """Fallback dummy embeddings for structure."""
    def embed_documents(self, texts: list[str]) -> list[list[float]]:
        return [[0.1] * 384 for _ in texts]
        
    def embed_query(self, text: str) -> list[float]:
        return [0.1] * 384

def get_qdrant_vector_store(collection_name: str = "rag_chunks") -> QdrantVectorStore:
    api_key = settings.QDRANT_API_KEY if settings.QDRANT_API_KEY else None
    client = QdrantClient(url=settings.QDRANT_URL, api_key=api_key)
    
    # Ideally, configure a real embedding model here
    embeddings = DummyEmbeddings()
    
    # Auto-create collection if it doesn't exist
    from qdrant_client.http import models
    if not client.collection_exists(collection_name):
        client.create_collection(
            collection_name=collection_name,
            vectors_config=models.VectorParams(size=384, distance=models.Distance.COSINE),
        )
    
    vector_store = QdrantVectorStore(
        client=client,
        collection_name=collection_name,
        embedding=embeddings,
    )
    return vector_store
